\name{WCluster-package}
\alias{WCluster-package}

\title{Clustering and PCA with Observational Weights and/or Variable Weights, and Data Nuggets Clustering
}

\description{
This package contains functions for K-means clustering, hierarchical clustering, and PCA with observational weights and/or variable weights. It also includes the corresponding functions for data nuggets which serve as representative samples of large datasets.
}

\author{
Yajie Duan, Javier Cabrera, Ge Cheng
}

\references{
Amaratunga, D., & Cabrera, J. (2009). Exploration and analysis of DNA microarray and protein array data. \emph{John Wiley & Sons} (Vol. 605).

Cherasia, K. E., Cabrera, J., Fernholz, L. T., & Fernholz, R. (2022). Data Nuggets in Supervised Learning. \emph{In Robust and Multivariate Statistical Methods: Festschrift in Honor of David E. Tyler} (pp. 429-449). Cham: Springer International Publishing.

Beavers, T., Cheng, G., Duan, Y., Cabrera, J., Lubomirski, M., Amaratunga, D., Teigler, J. (2023). Data Nuggets: A Method for Reducing Big Data While Preserving Data Structure (Submitted for Publication)
}
\seealso{
\code{\link[datanugget]{datanugget-package}}, \code{\link{Wkmeans}}, \code{\link{Whclust}}, \code{\link{Wpca}},  \code{\link{DN.Wkmeans}}
}
